--- Packages/com.unity.inputsystem/InputSystem/Plugins/UI/UIInputModule.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 597ec82a897364d78afe6ac0ea73518d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
